server/requirements.txt
pandas>=2.0.0
openai>=1.0.0
instructor>=0.4.0
pydantic>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
chromadb>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0 